renamed config.py to constants.py
